updated topic echo
